third_party/blink/web_tests/shapedetection/detection-HTMLVideoElement-invalid-state.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

// Detector's detect() rejects on a HAVE_NOTHING HTMLVideoElement.
promise_test(async t => {
  const video = document.createElement("video");
  video.src = "";
  const videoWatcher = new EventWatcher(t, video, ["play", "error"]);
  video.load();
  await videoWatcher.wait_for("error");
  assert_equals(video.readyState, video.HAVE_NOTHING);
  const detector = new TextDetector();
  await promise_rejects(t, 'InvalidStateError', detector.detect(video));
}, "Text - detect(HTMLVideoElement) - HAVE_NOTHING");

// Detector's detect() rejects on a HAVE_METADATA HTMLVideoElement.
promise_test(async t => {
  const video = document.createElement("video");
  video.src = "../external/wpt/media/white.webm";
  video.loop = true;
  video.autoplay = true;
  const videoWatcher = new EventWatcher(t, video, ["loadedmetadata", "error"]);
  video.load();
  await videoWatcher.wait_for("loadedmetadata");
  assert_equals(video.readyState, video.HAVE_METADATA);
  const detector = new TextDetector();
  await promise_rejects(t, 'InvalidStateError', detector.detect(video));
}, "Text - detect(HTMLVideoElement) - HAVE_METADATA");

</script>
